pyproject.toml
[tool.poetry]
name = "recolecta"
version = "0.1.0"
description = ""
authors = ["Jose B <jose.zdy@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django-extensions = "^2.2.8"
django-bootstrap4 = "^1.1.1"
pillow = "^7.0.0"
django = "^3.0.3"
django-optimized-image = "^0.2.0"
django-multiselectfield = "^0.1.12"

[tool.poetry.dev-dependencies]
ipdb = "^0.12.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
